ADDED TO P5
some of the lines were very close to intersecting, and in fact looked
like they were. But since they were still one or two pixels off, no
circle was drawn. It sometimes appeared that the program wasn’t picking
up on all intersections when it wasn’t the case at all, so I cheated a
little by making the circles draw as long as it’s close enough to an
intersection